Hw04 auth

+++ model/contacts/removeContact.js
@@ -1,7 +1,8 @@
 const { Contact } = require('../../db')
 
-const removeContact = async contactId => {
-  return await Contact.findOneAndRemove({ _id: contactId })
+const removeContact = async (contactId, owner) => {
+  return await Contact.findOneAndRemove({ _id: contactId, owner })
+  // return await Contact.findByIdAndRemove(contactId)
 }
 
 module.exports = removeContact
